+++ lib/components/OverlayTrigger/hooks/useCloseByEsc.tsx
@@ -1,0 +1,18 @@
+import {useStore} from "../../../utils";
+import {useEffect} from "react";
+import {KEY} from "../../../constants";
+import {Type} from "../../../types";
+
+export function useCloseByEsc() {
+    const {bus} = useStore()
+
+    useEffect(() => {
+        const handle = (event: KeyboardEvent) => {
+            if (event.key === KEY.ESC)
+                bus.send(Type.ClearTrigger)
+        }
+
+        window.addEventListener('keydown', handle)
+        return () => window.removeEventListener('keydown', handle)
+    }, [])
+}
