+++ lib/components/OverlayTrigger/index.ts
@@ -1,1 +1,3 @@
 export * from "./OverlayTrigger"
+export {useOverlayProps} from "./hooks/useOverlayProps";
+export {useHandlerSelect} from "./hooks/useHandlerSelect";
