+++ lib/components/Whisper/hooks/useTrigger.tsx
@@ -1,8 +1,7 @@
 import {useEffect, useState} from "react";
-import {useStore} from "../../utils";
-import {Trigger, Type} from "../../types";
-import {useSelectionChangeListener} from "./useSelectionChangeListener";
-import {TriggerFinder} from "../../utils/TriggerFinder";
+import {useStore} from "../../../utils";
+import {Trigger, Type} from "../../../types";
+import {TriggerFinder} from "../../../utils/TriggerFinder";
 
 export const useTrigger = (): Trigger | undefined => {
     const [trigger, setTrigger] = useState<Trigger | undefined>()
@@ -10,8 +9,6 @@ export const useTrigger = (): Trigger | undefined => {
 
     useEffect(() => bus.listen(Type.ClearTrigger, _ => setTrigger(undefined)), [])
     useEffect(() => bus.listen(Type.CheckTrigger, _ => setTrigger(TriggerFinder.find(options))), [options])
-
-    useSelectionChangeListener()
 
     return trigger
 }
